created sv_seed_funcs, and compacting code
