+++ nodes/vector/variable_lacunarity.py
@@ -25,6 +25,7 @@ from mathutils import noise
 
 from sverchok.node_tree import SverchCustomTreeNode
 from sverchok.data_structure import (updateNode, Vector_degenerate)
+from sverchok.utils.sv_seed_funcs import get_offset, seed_adjusted
 
 # noise nodes
 # from http://www.blender.org/documentation/blender_python_api_current/mathutils.noise.html
@@ -46,23 +47,6 @@ noise_options = [
 
 def var_func(position, distortion, _noise_type1, _noise_type2):
     return noise.variable_lacunarity(position, distortion, _noise_type1, _noise_type2)
-
-
-def get_offset(seed):
-    if seed == 0:
-        offset = [0.0, 0.0, 0.0]
-    else:
-        noise.seed_set(seed)
-        offset = noise.random_unit_vector() * 10.0
-    return offset
-
-
-def seed_adjusted(vert_list, seed):
-    if seed == 0.0:
-        return vert_list
-
-    ox = get_offset(seed)
-    return [[v[0] + ox[0], v[1] + ox[1], v[2] + ox[2]] for v in vert_list]
 
 
 noise_dict = {t[0]: t[1] for t in noise_options}
@@ -96,7 +80,7 @@ class SvLacunarityNode(bpy.types.Node, SverchCustomTreeNode):
         self.inputs.new('VerticesSocket', 'Vertices')
         self.inputs.new('StringsSocket', 'Seed').prop_name = 'seed'
         self.inputs.new('StringsSocket', 'Distrortion').prop_name = 'distortion'
-        self.outputs.new('VerticesSocket', 'Noise S')
+        self.outputs.new('StringsSocket', 'Noise S')
 
     def draw_buttons(self, context, layout):
         layout.prop(self, 'noise_type1', text="Type")
